Add administrator permissions (#2082)
* Clarify ability.rb roles and extract other administrators branch

* Add AdministratorPermissions matrix and Administrator#can?

* Can / can't update_application_rejected

* Can / can't update_user_info

In order to do that we must move the Admin::UsersController#update action to a new Admin::JobApplication::UsersController#update one, so we can check the job_application state when checking administrator's permissions.

* Can / can't update_application_state

* Remove useless custom validation cant_skip_state

As authorised step transitions are now somehow hard coded in
administrator_permissions.yml, this validation becomes useless.

* Managing job applications is now done by administrator permissions

* Review: no need for instance variables

* Refactoring: simplify with business method

* Review: wordings

* Remove useless admin permissions (#2087)

* Remove JobApplicationFile ability (#2090)

Permissions on files are now managed by validate_by_* booleans on
JobApplicationFile model. We don't need this ability anymore.

* Can / can't update_application_preselection and update states
permissions

+++ app/javascript/js/board.js
@@ -47,6 +47,14 @@ export function boardManagement() {
 
                 if (newState === "rejected") {
                   var rejectionUrl = item.getAttribute("data-rejection-url");
+                  if (rejectionUrl === "#") {
+                    Snackbar.show({
+                      text: "Vous n'êtes pas autorisé(e) à refuser cette candidature.",
+                      showAction: false,
+                    });
+                    boardRedraw();
+                    return;
+                  }
                   Rails.ajax({
                     type: "GET",
                     url: rejectionUrl,
@@ -82,6 +90,14 @@ export function boardManagement() {
                 } else if (item.getAttribute("data-rejected") === "true") {
                   if (newState === "initial") {
                     var rejectionUrl = item.getAttribute("data-rejection-url");
+                    if (rejectionUrl === "#") {
+                      Snackbar.show({
+                        text: "Vous n'êtes pas autorisé(e) à annuler le refus de cette candidature.",
+                        showAction: false,
+                      });
+                      boardRedraw();
+                      return;
+                    }
                     var destroyUrl = rejectionUrl.replace("/new", "");
                     Rails.ajax({
                       type: "DELETE",
